+++ registry/client.go
@@ -57,16 +57,16 @@ func NewClient(services *disco.Disco, client *http.Client) *Client {
 	}
 }
 
-// Discover qeuries the host, and returns the url for the registry.
-func (c *Client) Discover(host svchost.Hostname) *url.URL {
-	service := c.services.DiscoverServiceURL(host, serviceID)
-	if service == nil {
-		return nil
+// Discover queries the host, and returns the url for the registry.
+func (c *Client) Discover(host svchost.Hostname) (*url.URL, error) {
+	service, err := c.services.DiscoverServiceURL(host, serviceID)
+	if err != nil {
+		return nil, err
 	}
 	if !strings.HasSuffix(service.Path, "/") {
 		service.Path += "/"
 	}
-	return service
+	return service, nil
 }
 
 // Versions queries the registry for a module, and returns the available versions.
@@ -76,9 +76,9 @@ func (c *Client) Versions(module *regsrc.Module) (*response.ModuleVersions, erro
 		return nil, err
 	}
 
-	service := c.Discover(host)
-	if service == nil {
-		return nil, fmt.Errorf("host %s does not provide Terraform modules", host)
+	service, err := c.Discover(host)
+	if err != nil {
+		return nil, err
 	}
 
 	p, err := url.Parse(path.Join(module.Module(), "versions"))
@@ -149,9 +149,9 @@ func (c *Client) Location(module *regsrc.Module, version string) (string, error)
 		return "", err
 	}
 
-	service := c.Discover(host)
-	if service == nil {
-		return "", fmt.Errorf("host %s does not provide Terraform modules", host.ForDisplay())
+	service, err := c.Discover(host)
+	if err != nil {
+		return "", err
 	}
 
 	var p *url.URL